move darkTheme to reearthTheme
