+++ web/src/services/theme/index.ts
@@ -2,12 +2,12 @@ export * from "./reearthTheme/common";
 export * from "./styled";
 export { default as fonts } from "./reearthTheme/common/fonts";
 export { default as metrics, metricsSizes } from "./reearthTheme/common/metrics";
-export { default } from "./darkTheme";
+export { default } from "./reearthTheme/darkTheme";
 export { default as Provider } from "./provider";
 export { default as PublishedAppProvider } from "./publishedAppProvider";
 export { usePublishTheme, mask, type SceneThemeOptions } from "./publishTheme";
 export { default as GlobalStyles } from "./reearthTheme/common/globalStyles";
 export { default as lightTheme } from "./lightTheme";
-export { default as darkTheme } from "./darkTheme";
+export { default as darkTheme } from "./reearthTheme/darkTheme";
 
 export type { PublishTheme } from "./types";
